requirements.txt
flask
spacy
pyyaml